rename ParquetProjection related struct
